requirements.txt
dashscope==1.13.4
ntchat==0.1.20
openai==0.28.0
openai_whisper==20231117
pydantic==1.10.12
Requests==2.31.0
